+++ CHANGES.rst
@@ -1,6 +1,18 @@
 CHANGELOG
 ---------
 
+0.8.0 (in preparation)
+~~~~~~~~~~~~~~~~~~~~~~
+
+- remove the '--filters' parameter for good
+- change config format to YAML
+- change default config path to $XDG_CONFIG_HOME/udiskie/config.yml
+- separate ignore filters from mount option filters
+- allow to match multiple attributes against a device (AND-wise)
+- allow to overwrite udiskies default handleability settings
+- raise exception if --config file doesn't exist
+
+
 0.7.0
 ~~~~~
 
@@ -26,6 +38,7 @@ There are some backward incompatible changes, hence the version break:
 - pass only device name to external password prompt
 - add '--quiet' command line option
 - ignore devices ignored by udev rules
+
 
 0.6.4
 ~~~~~